chore: Factor out common logic in workerd tests
There's a LOT of repeated junk in the various test worker.py files, this factors
out some of it into a common `TestRunnerEntrypoint`. I also tried to tidy up as
I did the refactoring.
