+++ packages/django-cf/tests/in_worker/worker/pyproject.toml
@@ -7,4 +7,8 @@ dependencies = [
     "pytest",
     "pytest-asyncio<1.2.0",
     "sqlparse",
+    "testlib",
 ]
+
+[tool.uv.sources]
+testlib = { path = "../testlib/dist/testlib-0.0.0-py3-none-any.whl" }
