+++ packages/django-cf/tests/in_worker/worker/src/worker.py
@@ -1,9 +1,5 @@
 # pyright: reportMissingImports=false
 
-import asyncio
-import contextlib
-import importlib.util
-import io
 import os
 from pathlib import Path
 from urllib.parse import urlparse
@@ -14,21 +10,12 @@ import pytest
 from _django_app import R2_LOCATION, django_wsgi_app
 from django.http import HttpResponse, JsonResponse, StreamingHttpResponse
 from django.urls import path
-from pyodide.webloop import WebLoop
+from testlib.entrypoint import TestRunnerEntrypoint
 from worker_durable_object import TestDurableObject  # noqa: F401
-from workers import Response, WorkerEntrypoint
 
 BASE_DIR = Path(__file__).parent
 os.environ.setdefault("DJANGO_ALLOW_ASYNC_UNSAFE", "true")
 
-
-async def _noop(*args):
-    pass
-
-
-# pytest-asyncio relies on these methods, which older Pyodide WebLoops omit.
-WebLoop.shutdown_asyncgens = _noop
-WebLoop.shutdown_default_executor = _noop
 
 if not django.conf.settings.configured:
     django.conf.settings.configure(
@@ -149,61 +136,6 @@ urlpatterns = [
 ]
 
 
-class ResultCollector:
-    def __init__(self):
-        self.results = {}
-
-    @staticmethod
-    def _key(item):
-        normalized = []
-        if item.cls is not None:
-            normalized.append(item.cls.__name__)
-        name = getattr(item, "originalname", None) or item.name
-        normalized.append(name[len("test_") :] if name.startswith("test_") else name)
-        return "__".join(normalized)
-
-    @pytest.hookimpl(hookwrapper=True)
-    def pytest_runtest_makereport(self, item, call):
-        outcome = yield
-        report = outcome.get_result()
-        key = self._key(item)
-
-        if report.when == "call":
-            if report.passed:
-                self.results[key] = {"status": "passed"}
-            elif report.skipped:
-                self.results[key] = {
-                    "status": "skipped",
-                    "reason": str(report.longrepr),
-                }
-            elif report.failed:
-                excinfo = call.excinfo
-                if excinfo is not None and excinfo.errisinstance(AssertionError):
-                    self.results[key] = {
-                        "status": "failed",
-                        "error": str(excinfo.value),
-                    }
-                else:
-                    self.results[key] = {
-                        "status": "error",
-                        "error": f"{excinfo.typename}: {excinfo.value}"
-                        if excinfo is not None
-                        else "unknown error",
-                        "traceback": report.longreprtext,
-                    }
-        elif report.when in ("setup", "teardown") and report.skipped:
-            self.results[key] = {
-                "status": "skipped",
-                "reason": str(report.longrepr),
-            }
-        elif report.when in ("setup", "teardown") and report.failed:
-            self.results[key] = {
-                "status": "error",
-                "error": report.longreprtext,
-                "traceback": report.longreprtext,
-            }
-
-
 class EnvPlugin:
     def __init__(self, env):
         self._env = env
@@ -213,50 +145,12 @@ class EnvPlugin:
         return self._env
 
 
-class Default(DjangoCF, WorkerEntrypoint):
+class Default(DjangoCF, TestRunnerEntrypoint):
     def get_app(self):
         return django_wsgi_app()
 
     async def fetch(self, request):
         path = urlparse(request.url).path
-
-        if path.startswith("/run-tests/"):
-            suite_name = path[len("/run-tests/") :]
-            return self._run_suite(suite_name)
-        if path == "/health":
-            return Response.json({"ok": True})
         if path.startswith("/django/"):
-            return await super().fetch(request)
-        return Response.json({"error": "not found"}, status=404)
-
-    def _run_suite(self, suite_name):
-        module = f"test_{suite_name}"
-        if importlib.util.find_spec(module) is None:
-            return Response.json(
-                {"error": f"Unknown suite '{suite_name}' (no module '{module}')"},
-                status=404,
-            )
-
-        collector = ResultCollector()
-        saved_loop = asyncio.events._get_running_loop()
-        output = io.StringIO()
-        try:
-            with contextlib.redirect_stdout(output), contextlib.redirect_stderr(output):
-                exit_code = pytest.main(
-                    ["--pyargs", module, "-p", "no:cacheprovider"],
-                    plugins=[collector, EnvPlugin(self.env)],
-                )
-        finally:
-            asyncio.events._set_running_loop(saved_loop)
-        if exit_code != 0 and not collector.results:
-            return Response.json(
-                {
-                    "__session__": {
-                        "status": "error",
-                        "error": f"pytest exit code {exit_code}",
-                        "traceback": output.getvalue(),
-                    }
-                },
-                status=500,
-            )
-        return Response.json(collector.results)
+            return await DjangoCF.fetch(self, request)
+        return await TestRunnerEntrypoint.fetch(self, request)
