+++ packages/runtime-sdk/tests/bindings-test/src/worker.py
@@ -8,13 +8,7 @@ test_bindings.py maps onto individual pytest cases.
 To add a new binding: create `src/test_<binding>.py` with pytest tests.
 """
 
-import asyncio
-import importlib.util
-import sys
-from asyncio import InvalidStateError
-
-import pytest
-from pyodide.webloop import WebLoop
+from testlib.entrypoint import TestRunnerEntrypoint
 from worker_durable_object import (
     TestDurableObject,  # noqa: F401 - import to trigger side effect of registering the Durable Object
 )
@@ -22,132 +16,10 @@ from worker_workflow import (
     TestWorkflow,  # noqa: F401 - import to trigger side effect of registering the Workflow
 )
 
-from workers import Response, WorkerEntrypoint
-
-
-async def _noop(*args):
-    pass
-
-
-# pytest-asyncio relies on these but in Pyodide < 0.29 WebLoop does not implement them.
-WebLoop.shutdown_asyncgens = _noop
-WebLoop.shutdown_default_executor = _noop
-
-# Pyodide 0.26.0a2's WebLoop causes InvalidStateError when the
-# _cancel_all_tasks calls task.exception() on done-but-not-cancelled tasks.
-# Replace with a version that cancels tasks but tolerates that error.
-if sys.version_info < (3, 13):
-
-    def _cancel_all_tasks(loop):
-        to_cancel = asyncio.tasks.all_tasks(loop)
-        if not to_cancel:
-            return
-        for task in to_cancel:
-            task.cancel()
-        loop.run_until_complete(
-            asyncio.tasks.gather(*to_cancel, return_exceptions=True)
-        )
-        for task in to_cancel:
-            if task.cancelled():
-                continue
-            try:
-                if task.exception() is not None:
-                    loop.call_exception_handler(
-                        {
-                            "message": "unhandled exception during asyncio.run() shutdown",
-                            "exception": task.exception(),
-                            "task": task,
-                        }
-                    )
-            # Note: This exception catch is added from the original implementation
-            except (InvalidStateError, RuntimeError):
-                pass
-
-    asyncio.runners._cancel_all_tasks = _cancel_all_tasks  # type: ignore[attr-defined]
-
-
-class ResultCollector:
-    """pytest plugin that records each test's outcome keyed by its short name.
-
-    The "test_" prefix is stripped so keys match the names registered in
-    tests/test_bindings.py (e.g. test_put_and_get -> "put_and_get").
-    """
-
-    def __init__(self):
-        self.results = {}
-
-    @staticmethod
-    def _key(item):
-        name = item.name
-        return name[len("test_") :] if name.startswith("test_") else name
-
-    @pytest.hookimpl(hookwrapper=True)
-    def pytest_runtest_makereport(self, item, call):
-        outcome = yield
-        report = outcome.get_result()
-        key = self._key(item)
-
-        if report.when == "call":
-            if report.passed:
-                self.results[key] = {"status": "passed"}
-            elif report.skipped:
-                self.results[key] = {
-                    "status": "skipped",
-                    "reason": str(report.longrepr),
-                }
-            elif report.failed:
-                excinfo = call.excinfo
-                if excinfo is not None and excinfo.errisinstance(AssertionError):
-                    self.results[key] = {
-                        "status": "failed",
-                        "error": str(excinfo.value),
-                    }
-                else:
-                    self.results[key] = {
-                        "status": "error",
-                        "error": f"{excinfo.typename}: {excinfo.value}"
-                        if excinfo is not None
-                        else "unknown error",
-                        "traceback": report.longreprtext,
-                    }
-        elif report.when in ("setup", "teardown") and report.skipped:
-            self.results[key] = {
-                "status": "skipped",
-                "reason": str(report.longrepr),
-            }
-        elif report.when in ("setup", "teardown") and report.failed:
-            self.results[key] = {
-                "status": "error",
-                "error": report.longreprtext,
-                "traceback": report.longreprtext,
-            }
-
-
-class EnvPlugin:
-    def __init__(self, env):
-        self._env = env
-
-    @pytest.fixture
-    def env(self):
-        return self._env
-
-
 RECEIVED_MESSAGES = []
 
 
-class Default(WorkerEntrypoint):
-    async def fetch(self, request):
-        from urllib.parse import urlparse
-
-        path = urlparse(request.url).path
-
-        if path.startswith("/run-tests/"):
-            suite_name = path[len("/run-tests/") :]
-            return self._run_suite(suite_name)
-        if path == "/health":
-            return Response.json({"ok": True})
-        return Response.json({"error": "not found"}, status=404)
-
+class Default(TestRunnerEntrypoint):
     async def queue(self, batch, env, ctx):
         for message in batch.messages:
             RECEIVED_MESSAGES.append(
@@ -158,29 +30,3 @@ class Default(WorkerEntrypoint):
                 }
             )
             message.ack()
-
-    def _run_suite(self, suite_name):
-        module = f"test_{suite_name}"
-        if importlib.util.find_spec(module) is None:
-            return Response.json(
-                {"error": f"Unknown suite '{suite_name}' (no module '{module}')"},
-                status=404,
-            )
-
-        collector = ResultCollector()
-        # pytest-asyncio drives each test through asyncio.Runner, which calls
-        # asyncio.new_event_loop(). In Pyodide that constructs a WebLoop whose
-        # __init__ calls asyncio._set_running_loop(self) and is never restored on
-        # close(), so after pytest.main() the running loop points at an abandoned
-        # WebLoop. Save and restore it so the next request's fetch coroutine runs
-        # on the real workerd-driven loop instead of hanging on a dead one.
-        # TODO: fix this behavior in Pyodide
-        saved_loop = asyncio.events._get_running_loop()
-        try:
-            pytest.main(
-                ["--pyargs", module, "-p", "no:cacheprovider"],
-                plugins=[collector, EnvPlugin(self.env)],
-            )
-        finally:
-            asyncio.events._set_running_loop(saved_loop)
-        return Response.json(collector.results)
