+++ packages/runtime-sdk/tests/web-frameworks-test/django-async/pyproject.toml
@@ -10,4 +10,8 @@ dependencies = [
     "django-cors-headers",
     "pytest",
     "pytest-asyncio<1.2.0",
+    "testlib",
 ]
+
+[tool.uv.sources]
+testlib = { path = "../testlib/dist/testlib-0.0.0-py3-none-any.whl" }
