+++ packages/runtime-sdk/tests/web-frameworks-test/fastapi-tests/src/worker.py
@@ -1,7 +1,6 @@
 import asyncio
 import enum
 import hashlib
-import importlib.util
 import json
 from contextlib import asynccontextmanager
 from pathlib import Path
@@ -34,6 +33,7 @@ from pydantic import BaseModel
 from pyodide.webloop import WebLoop
 from starlette.background import BackgroundTask
 from starlette.middleware.gzip import GZipMiddleware
+from testlib.entrypoint import TestRunner
 
 import asgi
 
@@ -585,57 +585,6 @@ async def native_file():
 app.mount("/static", StaticFiles(directory=STATIC_DIR), name="static")
 
 
-class ResultCollector:
-    def __init__(self):
-        self.results = {}
-
-    @staticmethod
-    def _key(item):
-        name = item.name
-        return name[len("test_") :] if name.startswith("test_") else name
-
-    @pytest.hookimpl(hookwrapper=True)
-    def pytest_runtest_makereport(self, item, call):
-        outcome = yield
-        report = outcome.get_result()
-        key = self._key(item)
-
-        if report.when == "call":
-            if report.passed:
-                self.results[key] = {"status": "passed"}
-            elif report.skipped:
-                self.results[key] = {
-                    "status": "skipped",
-                    "reason": str(report.longrepr),
-                }
-            elif report.failed:
-                excinfo = call.excinfo
-                if excinfo is not None and excinfo.errisinstance(AssertionError):
-                    self.results[key] = {
-                        "status": "failed",
-                        "error": str(excinfo.value),
-                    }
-                else:
-                    self.results[key] = {
-                        "status": "error",
-                        "error": f"{excinfo.typename}: {excinfo.value}"
-                        if excinfo is not None
-                        else "unknown error",
-                        "traceback": report.longreprtext,
-                    }
-        elif report.when in ("setup", "teardown") and report.skipped:
-            self.results[key] = {
-                "status": "skipped",
-                "reason": str(report.longrepr),
-            }
-        elif report.when in ("setup", "teardown") and report.failed:
-            self.results[key] = {
-                "status": "error",
-                "error": report.longreprtext,
-                "traceback": report.longreprtext,
-            }
-
-
 class EnvPlugin:
     def __init__(self, env):
         self._env = env
@@ -653,27 +602,9 @@ class FastAPIAppPlugin:
 
 @app.get("/run-tests/{suite_name:path}")
 async def run_suite(suite_name: str, request: Request):
-    module = f"test_{suite_name}"
-    if importlib.util.find_spec(module) is None:
-        return JSONResponse(
-            {"error": f"Unknown suite '{suite_name}' (no module '{module}')"},
-            status_code=404,
-        )
-
-    collector = ResultCollector()
-    saved_loop = asyncio.events._get_running_loop()
-    try:
-        pytest.main(
-            ["--pyargs", module, "-p", "no:cacheprovider"],
-            plugins=[
-                collector,
-                EnvPlugin(request.scope["env"]),
-                FastAPIAppPlugin(),
-            ],
-        )
-    finally:
-        asyncio.events._set_running_loop(saved_loop)
-    return collector.results
+    runner = TestRunner(request.scope["env"])
+    result = runner.run_suite(suite_name)
+    return JSONResponse(result.payload, status_code=result.status)
 
 
 @app.get("/{path:path}")
