+++ packages/testlib/README.md
@@ -1,0 +1,1 @@
+These are testing helpers intended for use by the workers-py test suite.
