+++ packages/testlib/pyproject.toml
@@ -1,0 +1,9 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "testlib"
+version = "0.0.0"
+requires-python = ">=3.12"
+dependencies = ["pytest"]
